Data Analysis on Fifa_2019 Dataset
Data used for the particular analysis was downloaded from the kaggle. csv file is also attached. In this analysis we have made a team which contains best player for every position.  That team consist of a 4-3-3 football team formation. Data cleaning and pre-processing was also done before analysing. Visualization is done for better understanding.